+++ apps/designer/app/routes/auth/google/callback.tsx
@@ -1,6 +1,7 @@
 import type { LoaderArgs } from "@remix-run/node";
 import { authenticator } from "~/services/auth.server";
-import { returnToPath, loginPath } from "~/shared/router-utils";
+import { loginPath } from "~/shared/router-utils";
+import { returnToPath } from "~/services/cookie.server";
 
 export const loader = async ({ request }: LoaderArgs) => {
   const returnTo = await returnToPath(request);
