+++ apps/designer/app/services/cookie.server.ts
@@ -1,4 +1,5 @@
 import { createCookie } from "@remix-run/node";
+import { dashboardPath } from "~/shared/router-utils";
 
 export const returnToCookie = createCookie("returnTo", {
   path: "/",
@@ -7,3 +8,11 @@ export const returnToCookie = createCookie("returnTo", {
   maxAge: 60, // 1 minute because it makes no sense to keep it for a long time
   secure: process.env.NODE_ENV === "production",
 });
+
+export const returnToPath = async (request: Request) => {
+  const returnTo =
+    (await returnToCookie.parse(request.headers.get("Cookie"))) ??
+    dashboardPath();
+
+  return returnTo;
+};
